requirements.txt
requests>=2.31.0
click>=8.1.7
python-dotenv>=1.0.0
openai>=1.6.0
anthropic>=0.18.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
rich>=13.7.0